requirements.txt
fastmcp
python-docx
fpdf2
openpyxl
pdfplumber
Pillow
matplotlib
